scripts/requirements_anomaly.txt
# Requirements for DINOv3 Anomaly Detection Project

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace for pretrained models
transformers>=4.56.0

# Data processing
numpy>=1.24.0
Pillow>=9.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Progress bars
tqdm>=4.65.0

# Optional: For better performance
# accelerate>=0.20.0  # For faster HuggingFace model loading
